+++ R/brmsterms.R
@@ -532,13 +532,7 @@ terms_resp <- function(formula, check_names = TRUE) {
   } else {
     str_fun <- deparse_no_string(expr[[1]]) 
     use_mvbind <- identical(str_fun, "mvbind")
-    use_cbind <- identical(str_fun, "cbind")
     if (use_mvbind) {
-      out <- ulapply(expr[-1], deparse_no_string)
-    } else if (use_cbind) {
-      # deprecated as of brms 2.7.2
-      warning2("Using 'cbind' for multivariate models is ", 
-               "deprecated. Please use 'mvbind' instead.")
       out <- ulapply(expr[-1], deparse_no_string)
     } else {
       out <- deparse_no_string(expr)