+++ R/posterior_epred.R
@@ -145,6 +145,7 @@ posterior_epred.brmsprep <- function(object, scale, dpar, nlpar, sort,
         out <- posterior_epred_fun(object)
       }
     } else {
+      # return results on the linear scale
       if (conv_cats_dpars(object$family)) {
         mus <- dpars[grepl("^mu", dpars)] 
       } else {
@@ -273,14 +274,13 @@ fitted.brmsfit <- function(object, newdata = NULL, re_formula = NULL,
 #' 
 #' @inheritParams posterior_epred.brmsfit
 #' @param object An object of class \code{brmsfit}.
-#' @param transform (Deprecated) Logical; if \code{FALSE}
+#' @param transform Logical; if \code{FALSE}
 #'  (the default), samples of the linear predictor are returned.
 #'  If \code{TRUE}, samples of transformed linear predictor,
-#'  that is, the mean of the posterior predictive distribution
-#'  are returned instead (see \code{\link{posterior_epred}} for details).
-#'  Only implemented for compatibility with the 
-#'  \code{\link[rstantools:posterior_linpred]{posterior_linpred}}
-#'  generic. 
+#'  that is, after applying the link function are returned.
+#' @param dpar Name of a predicted distributional parameter
+#'  for which samples are to be returned. By default, samples
+#'  of the main distributional parameter(s) \code{"mu"} are returned.
 #' @param incl_thres For ordinal families only: A single logical value
 #'   indicating whether to take the thresholds (intercepts) into account
 #'   (\code{TRUE}) or not (\code{FALSE}). Thereby, "taking the thresholds into
@@ -318,9 +318,13 @@ posterior_linpred.brmsfit <- function(
   scale <- "linear"
   transform <- as_one_logical(transform)
   if (transform) {
-    warning2("posterior_linpred(transform = TRUE) is deprecated. Please ",
-             "use posterior_epred() instead, without the 'transform' argument.")
     scale <- "response"
+    # if transform, return inv-link samples of only a single
+    # distributional or non-linear parameter for consistency
+    # of brms and rstanarm
+    if (is.null(dpar) && is.null(nlpar)) {
+      dpar <- "mu"
+    }
   }
   contains_samples(object)
   object <- restructure(object)
@@ -534,6 +538,18 @@ posterior_epred_dirichlet <- function(prep) {
   out <- aperm(out, perm = c(1, 3, 2))
   dimnames(out)[[3]] <- prep$cats
   out
+}
+
+posterior_epred_dirichlet2 <- function(prep) {
+  mu <- prep$dpars[grepl("^mu", names(prep$dpars))]
+  mu <- abind(mu, along = 3)
+  sums_mu <- apply(mu, 1:2, sum)
+  cats <- seq_len(prep$data$ncat)
+  for (i in cats) {
+    mu[, , i] <- mu[, , i] / sums_mu
+  }
+  dimnames(mu)[[3]] <- prep$cats
+  mu
 }
 
 posterior_epred_cumulative <- function(prep) {