+++ R/stan-predictor.R
@@ -1941,7 +1941,7 @@ stan_eta_rsp <- function(r) {
 stan_eta_transform <- function(family, cens_or_trunc = FALSE) {
   transeta <- "transeta" %in% family_info(family, "specials")
   no_transform <- family$link == "identity" && !transeta || 
-    is_polytomous(family) && !is.customfamily(family)
+    has_joint_link(family) && !is.customfamily(family)
   !no_transform && !stan_has_built_in_fun(family, cens_or_trunc)
 }